+++ engine/engine/components/comp_camera.h
@@ -46,7 +46,7 @@ public:
 		assert(trans || fatal("TCamera requieres a TTransform component"));
 
 		trans->lookAt(target, trans->getUp());
-		setViewport(0, 0, CApp::get().xres, CApp::get().yres);
+		setViewport(0, 0, (float)CApp::get().xres, (float)CApp::get().yres);
 	}
 
 	void update(float elapsed) {
